1379: Retrieve client_id from request object JWT for PAR
client_id is extracted from the request JWT's claim of the same name.

The previous impl was incorrect, as it assumed that the client_id was always present as a form param. This is only the case due to OPENAM-21910 bug in AM. This impl will work when the bug is fixed in AM.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1379