1417: Add workaround for AM PAR endpoint bug requiring client_id param
An AM bug exists which requires the client_id param be supplied in all calls to the /par endpoint, see https://bugster.forgerock.org/jira/browse/OPENAM-21910

According to the spec, the client_id is only required when it is needed as an authentication param i.e. when the client is doing tls_client_auth. If the client is doing private_key_jwt auth then the client_id is not a required param.

This workaround uses a filter to add the client_id param if it is missing, the value is sourced from the request JWT param's iss claim.

The workaround is best-effort, if an error occurs in it is ignored and the call is allowed to pass through to AM.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1417