1517: Move JWKS locator logic into ApiClient
Refactor ApiClient to move the JWKS locator logic internal to this class.

The builder of an ApiClient can specify either the embedded JWKS or a jwks_uri value. Internally, a Supplier is created which returns a Promise<JWKSet, FailedToLoadJWKException>, this provides a consistent way to retrieve the JWKS and hides the implementation details inside the ApiClient.

Removed filters and services which previously implemented the logic to retrieve the JWKS based on whether the ApiClient.jwks or ApiClient.jwksUri fields were set.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1517